refactor: absorb topic_normalizer into topic_filter
Move SegmentFormat, to_canonical, to_mqtt, to_topic_filter, detect_format
and helpers into topic_filter.py — they are all part of the same topic-
format domain that EventTopicFilter owns.

- Delete axis/models/events/topic_normalizer.py
- Update three interface importers (event_instances, event_manager, mqtt)
- Update tests/test_topic_normalizer.py to import from topic_filter
- Export new symbols from axis/models/events/__init__.py

No behaviour change; all 463 tests pass, coverage 98.01%.
