+++ axis/models/events/__init__.py
@@ -1,5 +1,21 @@
 """Event-related models namespace."""
 
-from .topic_filter import EventTopicFilter, from_desired_subscriptions
+from .topic_filter import (
+    EventTopicFilter,
+    SegmentFormat,
+    detect_format,
+    from_desired_subscriptions,
+    to_canonical,
+    to_mqtt,
+    to_topic_filter,
+)
 
-__all__ = ["EventTopicFilter", "from_desired_subscriptions"]
+__all__ = [
+    "EventTopicFilter",
+    "SegmentFormat",
+    "detect_format",
+    "from_desired_subscriptions",
+    "to_canonical",
+    "to_mqtt",
+    "to_topic_filter",
+]
